backend/migration-script/gcp-to-aws/requirements.txt
cognitojwt==1.1.0
PyMySQL==1.0.2
requests==2.26.0
boto3==1.20.7
pytz==2021.3
python-dotenv==0.19.2
hl7apy==1.3.4
paramiko==2.11.0